Example 2
Using a do while loop, a boolean set to false and if else statement to alter the boolean. Ask user to input a double. Take the user input and check if the std::istringstream input can be saved into the double variable, if so we have an integer if not we display an error message.